requirements.txt
# Python dependencies

fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
python-jose==3.3.0
passlib==1.7.4
python-multipart==0.0.6
redis==5.0.1
apscheduler==3.10.4
python-dotenv==1.0.0
bcrypt==4.0.1
pydantic==2.5.2